pyproject.toml
[project]
name = "ai-lead-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "firebase-admin",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",
    "httpx",
    "beautifulsoup4",
]
